Update cib_parser.py
For specific resource types or global properties, they have a default tuned value. So even in the cib.xml they are not defined, it is considered as a good configuration. However, if they are explictly defined in cib.xml but conflicts with the predefined value in `cib_parameters_value` line, we need give a warning same as it does in existing code.